Add code to automatically scroll to bottom and move classes
Move the gui package to ui as it all belongs under user interface.
